compilation on Fedora Core 11
git-svn-id: file:///home/karl/SVN2/amilab.svn/code/trunk@210 52b0d089-85ac-4fb5-acf5-4e3172ea7aaf

+++ src/Wrapping/WrapVTK/vtkwrap.cpp
@@ -140,7 +140,7 @@ SurfacePoly* Func_decimate( SurfacePoly* surf, float target_reduction )
 
   surf_result = new SurfacePoly(deci->GetOutput());
 
-  printf("Resulting image: %d points \n",deci->GetOutput()->GetNumberOfPoints());
+  printf("Resulting image: %d points \n",(int)deci->GetOutput()->GetNumberOfPoints());
 
   triangle_filter  ->Delete();
   deci             ->Delete();
